Rewrite stringify
Completely rewrite the stringify process from scratch. This is to
address the many deficiencies with the current process, which is
fundamentally flawed due to the fact that it relies on stringifying to
JSON, which lacks specificity for functions, regexes, and more. I've
removed commas from object/array stringification as well, which should
make for cleaner diffs. Lastly, since the new process utilizes
Jasmine.pp for non-array/object values, the output should better
resemble default Jasmine pretty printing.